sheets_parsing/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
creds
psycopg2-binary==2.9.3
bs4
lxml
notifiers